requirements/local.txt
asgiref==3.2.7
Django==3.0.6
django-model-utils==4.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
Unipath==1.1